+++ click_extended/decorators/check/__init__.py
@@ -1,6 +1,7 @@
 """Initialization file for the `click_extended.decorators.check` module."""
 
 from click_extended.decorators.check.conflicts import conflicts
+from click_extended.decorators.check.contains import contains
 from click_extended.decorators.check.dependencies import dependencies
 from click_extended.decorators.check.ends_with import ends_with
 from click_extended.decorators.check.exclusive import exclusive
@@ -27,6 +28,7 @@ from click_extended.decorators.check.truthy import truthy
 
 __all__ = [
     "conflicts",
+    "contains",
     "dependencies",
     "ends_with",
     "exclusive",