+++ click_extended/decorators/transform/__init__.py
@@ -3,8 +3,12 @@
 from click_extended.decorators.transform.add_prefix import add_prefix
 from click_extended.decorators.transform.add_suffix import add_suffix
 from click_extended.decorators.transform.apply import apply
+from click_extended.decorators.transform.expand_vars import expand_vars
 from click_extended.decorators.transform.remove_prefix import remove_prefix
 from click_extended.decorators.transform.remove_suffix import remove_suffix
+from click_extended.decorators.transform.replace import replace
+from click_extended.decorators.transform.slugify import slugify
+from click_extended.decorators.transform.split import split
 from click_extended.decorators.transform.strip import lstrip, rstrip, strip
 from click_extended.decorators.transform.to_case import (
     to_camel_case,
@@ -26,15 +30,20 @@ from click_extended.decorators.transform.to_datetime import to_datetime
 from click_extended.decorators.transform.to_path import to_path
 from click_extended.decorators.transform.to_time import to_time
 from click_extended.decorators.transform.to_timestamp import to_timestamp
+from click_extended.decorators.transform.truncate import truncate
 
 __all__ = [
     "add_prefix",
     "add_suffix",
     "apply",
+    "expand_vars",
     "lstrip",
     "remove_prefix",
     "remove_suffix",
+    "replace",
     "rstrip",
+    "slugify",
+    "split",
     "strip",
     "to_camel_case",
     "to_dot_case",
@@ -54,4 +63,5 @@ __all__ = [
     "to_path",
     "to_time",
     "to_timestamp",
+    "truncate",
 ]